+++ qc-application/src/main/java/com/eager/questioncloud/application/api/cart/dto/CartRequest.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.cart;
+package com.eager.questioncloud.application.api.cart.dto;
 
 import java.util.List;
 import lombok.Getter;
